// data/github/current/ui/packages/memex/src/client/components/automation/__generated__/searchInputLabelSuggestionsPaginated.graphql.ts
export type searchInputLabelSuggestionsPaginated$data = {
  readonly id: string;
  readonly labels: {
    readonly edges: ReadonlyArray<{
      readonly __id: string;
      readonly node: {
        readonly " $fragmentSpreads": FragmentRefs<"searchInputLabel">;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  } | null | undefined;
  readonly " $fragmentType": "searchInputLabelSuggestionsPaginated";
};